Added Validation Methods
Added validation methods for CareInfo
